+++ airflow_indexima/hooks/indexima.py
@@ -1,11 +1,17 @@
 """Indexima hook module definition."""
 
-from typing import Any, List, Optional
+import datetime
+from typing import Any, Dict, List, Optional, Union
 
 from airflow.hooks.base_hook import BaseHook
 from pyhive import hive
 
-from airflow_indexima.connection import ConnectionDecorator
+from airflow_indexima.connection import (
+    ConnectionDecorator,
+    apply_hive_extra_setting,
+    extract_hive_extra_setting,
+)
+from airflow_indexima.hive_transport import create_hive_transport
 
 
 __all__ = ['IndeximaHook']
@@ -30,8 +36,12 @@ class IndeximaHook(BaseHook):
     def __init__(
         self,
         indexima_conn_id: str,
-        auth: str = 'CUSTOM',
         connection_decorator: Optional[ConnectionDecorator] = None,
+        dry_run: Optional[bool] = False,
+        auth: Optional[str] = None,
+        kerberos_service_name: Optional[str] = None,
+        timeout_seconds: Optional[Union[int, datetime.timedelta]] = None,
+        socket_keepalive: Optional[bool] = None,
         *args,
         **kwargs,
     ):
@@ -42,13 +52,42 @@ class IndeximaHook(BaseHook):
             auth(str): pyhive authentication mode (defaults: 'CUSTOM')
             connection_decorator (Optional[ConnectionDecorator]) : optional function handler
                 to post process connection parameter(default: None)
+            dry_run (Optional[bool]): dry run mode (default: False). If true no action will
+                be applied against datasource.
+            timeout_seconds (Optional[Union[int, datetime.timedelta]]): define the socket timeout in second
+                (could be an int or a timedelta)
+            socket_keepalive (Optional[bool]): enable TCP keepalive.
+            kerberos_service_name (Optional[str]): optional kerberos service name
+
+        Per default, hive connection is set in 'utf-8':
+        ```{ "serialization.encoding": "utf-8"}```
+
         """
         super(IndeximaHook, self).__init__(source='indexima', *args, **kwargs)
         self._indexima_conn_id = indexima_conn_id
         self._schema = kwargs.pop("schema", None)
-        self._auth = auth
+
         self._conn: Optional[Any] = None
+        self._cursor: Optional[Any] = None
         self._connection_decorator = connection_decorator
+        self._dry_run = dry_run or False
+
+        _timeout_seconds = None
+        if timeout_seconds is not None:
+            if isinstance(timeout_seconds, datetime.timedelta):
+                _timeout_seconds = timeout_seconds.seconds
+            else:
+                _timeout_seconds = int(timeout_seconds)
+
+        self._settings_decorator = lambda connection: apply_hive_extra_setting(
+            connection=connection,
+            auth=auth,
+            kerberos_service_name=kerberos_service_name,
+            timeout_seconds=_timeout_seconds,
+            socket_keepalive=socket_keepalive,
+        )
+        # set default hive configuration
+        self._hive_configuration: Optional[Dict[str, str]] = {"serialization.encoding": "utf-8"}
 
     def get_conn(self) -> hive.Connection:
         """Return a hive connection.
@@ -60,19 +99,38 @@ class IndeximaHook(BaseHook):
         conn = self.get_connection(self._indexima_conn_id)
         if not conn:
             raise RuntimeError(f'no connection identifier found with {self._indexima_conn_id}')
+
+        # load extra parameters of airflow connection
+        conn = self._settings_decorator(conn)
+
+        # apply decorator
         if self._connection_decorator:
             conn = self._connection_decorator(conn)
-        self.log.info(
-            f'connect to {conn.host}  {conn.login}  {"X"*len(conn.password)}  {conn.port} {self._auth}'  # noqa: E501
+
+        self.log.info(f'connect to {conn.host}  {conn.login} {conn.port}')  # noqa: E501
+        (auth, kerberos_service_name, timeout_seconds, socket_keepalive) = extract_hive_extra_setting(
+            connection=conn
         )
+
+        # build parameters for create_hive_transport and keep default value meaning
+        parameters = {'host': conn.host}
+        parameters['port'] = conn.port or 10000
+        parameters['timeout_seconds'] = timeout_seconds or 60
+        if socket_keepalive is not None:
+            parameters['socket_keepalive'] = socket_keepalive
+        parameters['auth'] = auth or 'CUSTOM'
+        if conn.login:
+            parameters['username'] = conn.login
+        if conn.password:
+            parameters['password'] = conn.password
+        if kerberos_service_name:
+            parameters['kerberos_service_name'] = kerberos_service_name
         self._conn = hive.Connection(
-            host=conn.host,
-            username=conn.login,
-            password=conn.password,
+            configuration=self._hive_configuration,
             database=self._schema or conn.schema,
-            port=conn.port if conn.port else 10000,
-            auth=self._auth,
+            thrift_transport=create_hive_transport(**parameters),
         )
+        self._cursor = self._conn.cursor()  # type: ignore
         return self._conn
 
     def get_records(self, sql: str) -> hive.Cursor:
@@ -87,11 +145,13 @@ class IndeximaHook(BaseHook):
 
     def run(self, sql: str) -> hive.Cursor:
         """Execute query and return curror."""
-        if not self._conn:
+        if not self._cursor:
             self.get_conn()
-        cursor = self._conn.cursor()  # type: ignore
-        cursor.execute(sql)
-        return cursor
+        if not self._dry_run:
+            self._cursor.execute(sql)  # type: ignore
+        else:
+            self.log.warn(sql)
+        return self._cursor
 
     def check_error_of_load_query(self, cursor: hive.Cursor):
         """Raise error if a load query fail.
@@ -127,12 +187,51 @@ class IndeximaHook(BaseHook):
         """
         self.run(f'ROLLBACK {tablename}')
 
+    def close(self):
+        """Close current connection."""
+        if self._conn:
+            self._conn.close()
+        self._conn = None
+        self._cursor = None
+
     def __enter__(self):
         self.get_conn()
         return self
 
     def __exit__(self, *exc):
-        if self._conn:
-            self._conn.close()
-        self._conn = None
+        self.close()
         return False
+
+    def is_dry_run(self) -> bool:
+        return self._dry_run
+
+    @property
+    def hive_configuration(self) -> Optional[Dict[str, str]]:
+        """Return hive configuration.
+
+        # Returns
+            (Dict[str, str]): A dictionary of Hive settings (functionally same as the `set` command)
+        """
+        return self._hive_configuration
+
+    @hive_configuration.setter
+    def hive_configuration(self, configuration: Optional[Dict[str, str]]):
+        """Set hive connection configuration.
+
+        # Parameters
+            configuration: A dictionary of Hive settings (functionally same as the `set` command)
+
+        # Example
+
+        ```python
+        hool.hive_configuration = {
+            "hive.server.read.socket.timeout": str(3600000),
+            "hive.server2.session.check.interval": str(3600000),
+            "hive.server2.idle.session.check.operation": "true",
+            "hive.server2.idle.operation.timeout": str(3600000 * 24),
+            "hive.server2.idle.session.timeout": str(3600000 * 24 * 3),
+        })
+        ```
+
+        """
+        self._hive_configuration = configuration